Move event test to defaultLoop
